Break up server.go into multiple files
